# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface ToggleReviewedRequest {
  reviewed: boolean;
}

export interface RedeemMagicLinkRequest {
  token: string;
}

export type SyncResultSource =
  (typeof SyncResultSource)[keyof typeof SyncResultSource];

export const SyncResultSource = {
  notion: "notion",
  mock: "mock",
} as const;

export interface SyncResult {
  synced: number;
  source: SyncResultSource;
}

export interface UploadedFile {
  filename: string;
  driveFileId: string;
  /** @nullable */
  driveFileUrl?: string | null;
  size: number;
  uploadedAt: string;
}

export type ChecklistItemStatus =
  (typeof ChecklistItemStatus)[keyof typeof ChecklistItemStatus];

export const ChecklistItemStatus = {
  needed: "needed",
  uploaded: "uploaded",
  reviewed: "reviewed",
} as const;

export interface ChecklistItem {
  id: string;
  title: string;
  description: string;
  category: string;
  whyWeNeedThis: string;
  status: ChecklistItemStatus;
  position: number;
  files: UploadedFile[];
}

export interface ChecklistItemDraft {
  title: string;
  description: string;
  category: string;
  whyWeNeedThis: string;
  /** @nullable */
  position?: number | null;
}

export interface ChecklistItemUpdate {
  title?: string;
  description?: string;
  category?: string;
  whyWeNeedThis?: string;
  position?: number;
}

export interface CreateClientRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  /** @nullable */
  priorReturnSummary?: string | null;
  /** @nullable */
  driveFolderId?: string | null;
  tags?: string[];
}

export type AdminClientSummaryStatus =
  (typeof AdminClientSummaryStatus)[keyof typeof AdminClientSummaryStatus];

export const AdminClientSummaryStatus = {
  draft: "draft",
  published: "published",
  archived: "archived",
  no_session: "no_session",
} as const;

export interface AdminClientSummary {
  id: string;
  name: string;
  email: string;
  tags: string[];
  hasDriveFolder: boolean;
  itemsTotal: number;
  itemsReceived: number;
  itemsReviewed: number;
  status: AdminClientSummaryStatus;
  /** @nullable */
  lastUploadAt?: string | null;
  /** @nullable */
  publishedAt?: string | null;
}

export type AdminClientDetailStatus =
  (typeof AdminClientDetailStatus)[keyof typeof AdminClientDetailStatus];

export const AdminClientDetailStatus = {
  draft: "draft",
  published: "published",
  archived: "archived",
} as const;

export interface AdminClientDetail {
  id: string;
  name: string;
  email: string;
  tags: string[];
  hasDriveFolder: boolean;
  /** @nullable */
  priorReturnSummary?: string | null;
  filingYear: number;
  status: AdminClientDetailStatus;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  magicLinkUrl?: string | null;
  /** @nullable */
  magicLinkExpiresAt?: string | null;
  items: ChecklistItem[];
}

export interface PublishResult {
  magicLinkUrl: string;
  expiresAt: string;
}

export type EmailSendResultMode =
  (typeof EmailSendResultMode)[keyof typeof EmailSendResultMode];

export const EmailSendResultMode = {
  gmail: "gmail",
  mock: "mock",
} as const;

export interface EmailSendResult {
  sent: boolean;
  recipient: string;
  mode: EmailSendResultMode;
  /** @nullable */
  previewBody?: string | null;
}

export interface AdminDashboardActivity {
  clientId: string;
  clientName: string;
  filename: string;
  uploadedAt: string;
}

export interface AdminDashboard {
  totalClients: number;
  publishedSessions: number;
  itemsNeeded: number;
  itemsReceived: number;
  itemsReviewed: number;
  recentActivity: AdminDashboardActivity[];
}

export type ClientChecklistItemViewStatus =
  (typeof ClientChecklistItemViewStatus)[keyof typeof ClientChecklistItemViewStatus];

export const ClientChecklistItemViewStatus = {
  needed: "needed",
  uploaded: "uploaded",
  reviewed: "reviewed",
} as const;

export interface ClientChecklistItemView {
  id: string;
  title: string;
  description: string;
  category: string;
  whyWeNeedThis: string;
  status: ClientChecklistItemViewStatus;
  files: UploadedFile[];
}

export type ClientChecklistViewMode =
  (typeof ClientChecklistViewMode)[keyof typeof ClientChecklistViewMode];

export const ClientChecklistViewMode = {
  client: "client",
  admin_preview: "admin_preview",
} as const;

export type ClientChecklistViewSessionStatus =
  (typeof ClientChecklistViewSessionStatus)[keyof typeof ClientChecklistViewSessionStatus];

export const ClientChecklistViewSessionStatus = {
  draft: "draft",
  published: "published",
  archived: "archived",
  no_session: "no_session",
} as const;

export interface ClientChecklistView {
  clientFirstName: string;
  clientFullName: string;
  filingYear: number;
  itemsTotal: number;
  itemsCompleted: number;
  items: ClientChecklistItemView[];
  mode: ClientChecklistViewMode;
  /** @nullable */
  adminDisplayName?: string | null;
  sessionStatus?: ClientChecklistViewSessionStatus;
}

export interface AdminPreviewSessionResult {
  previewUrl: string;
  expiresAt: string;
  clientName: string;
}

/**
 * Not authenticated
 */
export type UnauthorizedResponse = ApiError;

/**
 * Resource not found
 */
export type NotFoundResponse = ApiError;