# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /portal/client/session:
    get:
      operationId: getClientSession
      tags: [portal-client]
      summary: Get the current client session
      description: Returns the authenticated client's checklist based on their magic-link cookie.
      responses:
        "200":
          description: Client checklist view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientChecklistView"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /portal/client/exit-preview:
    post:
      operationId: exitClientPreview
      tags: [portal-client]
      summary: End an admin preview session
      description: Clears the admin-preview cookie and revokes the preview session row. Only valid when the active session is an admin preview.
      responses:
        "204":
          description: Preview ended
        "401":
          $ref: "#/components/responses/Unauthorized"
  /portal/client/redeem:
    post:
      operationId: redeemMagicLink
      tags: [portal-client]
      summary: Redeem a magic-link token
      description: Validates the token, sets a short-lived signed session cookie, and returns the client checklist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemMagicLinkRequest"
      responses:
        "200":
          description: Token accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientChecklistView"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RedeemMagicLinkRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
...
    UploadedFile:
      type: object
      required: [filename, driveFileId, size, uploadedAt]
      properties:
        filename:
          type: string
        driveFileId:
          type: string
        driveFileUrl:
          type: ["string", "null"]
        size:
          type: integer
        uploadedAt:
          type: string
          format: date-time
...
    ClientChecklistItemView:
      type: object
      required: [id, title, description, category, whyWeNeedThis, status, files]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        whyWeNeedThis:
          type: string
        status:
          type: string
          enum: [needed, uploaded, reviewed]
        files:
          type: array
          items:
            $ref: "#/components/schemas/UploadedFile"
    ClientChecklistView:
      type: object
      required:
        - clientFirstName
        - clientFullName
        - filingYear
        - itemsTotal
        - itemsCompleted
        - items
        - mode
      properties:
        clientFirstName:
          type: string
        clientFullName:
          type: string
        filingYear:
          type: integer
        itemsTotal:
          type: integer
        itemsCompleted:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/ClientChecklistItemView"
        mode:
          type: string
          enum: [client, admin_preview]
        adminDisplayName:
          type: ["string", "null"]
        sessionStatus:
          type: string
          enum: [draft, published, archived, no_session]
...